修改了机制，理论上更好了

+++ testip0.sh
@@ -5,6 +5,18 @@ if [ $chk_res -eq 0 ]
 then
 	rm ns.log
 else
+	cat ns.log | while read line
+	do
+		chk_res=`cat black.lst | grep $line`
+		if [ ! -z $chk_res ]
+		then
+			#echo !!!!black list!!!!
+			rm ns.log
+		fi
+	done
+fi
+if [ -f  ns.log ]
+then
 	mv ns.log 'ns'$1'.log'
 fi
 if [ ! -f 'ns'$1'.log' ]
